Fix: Reconnect to move module importing. #181
